+++ software/io/command_interface/control_target.cc
@@ -5,8 +5,7 @@
 #include "c64.h"
 #include "tape_recorder.h"
 #include "reu_preloader.h"
-#include "iec.h"
-#include "iec_printer.h"
+#include "iec_interface.h"
 #if U64
 #include "u64_config.h"
 #include "u64.h"
@@ -245,19 +244,7 @@ void ControlTarget :: parse_command(Message *command, Message **reply, Message *
                 data_message.message[offs++] = c1541_B->get_drive_power() ? 1 : 0;
                 data_message.length += 3;
             }
-            if (iec_if) { // seems to be always available
-                data_message.message[0]++;
-                data_message.message[offs++] = 0x0F;
-                data_message.message[offs++] = (uint8_t)iec_if->get_current_iec_address();
-                data_message.message[offs++] = iec_if->iec_enable;
-                data_message.length += 3;
-
-                data_message.message[0]++;
-                data_message.message[offs++] = 0x50;
-                data_message.message[offs++] = (uint8_t)iec_printer->get_current_printer_address();
-                data_message.message[offs++] = (uint8_t)iec_printer->is_enabled();
-                data_message.length += 3;
-            }
+            IecInterface::info(data_message, offs);
             break;
         }
