--- software/io/iec/iec.cc
@@ -1,1241 +1,0 @@
-#include "integer.h"
-extern "C" {
-	#include "dump_hex.h"
-    #include "small_printf.h"
-}
-#include "iec.h"
-#include "iec_channel.h"
-#include "iec_printer.h"
-#include "c64.h"
-#include "u64.h"
-#include "filemanager.h"
-#include "userinterface.h"
-#include "pattern.h"
-#include "command_intf.h"
-#include "endianness.h"
-#include "init_function.h"
-#include "json.h"
-
-#define MENU_IEC_ON          0xCA0E
-#define MENU_IEC_OFF         0xCA0F
-
-#define MENU_IEC_RESET       0xCA10
-#define MENU_IEC_TRACE_ON    0xCA11
-#define MENU_IEC_TRACE_OFF   0xCA12
-
-#define MENU_IEC_WARP_8      0xCA13
-#define MENU_IEC_WARP_9      0xCA14
-#define MENU_IEC_WARP_10     0xCA15
-#define MENU_IEC_WARP_11     0xCA16
-
-#define MENU_IEC_SET_DIR     0xCA17
-#define MENU_IEC_MASTER_1    0xCA18
-#define MENU_IEC_MASTER_2    0xCA19
-#define MENU_IEC_MASTER_3    0xCA1A
-#define MENU_IEC_MASTER_4    0xCA1B
-#define MENU_IEC_LOADDIR     0xCA1C
-#define MENU_IEC_LOADFIRST   0xCA1D
-#define MENU_READ_STATUS     0xCA1E
-#define MENU_SEND_COMMAND    0xCA1F
-#define MENU_IEC_FLUSH       0xCA20
-   
-//extern BYTE _warp_rom_65_start;
-extern uint8_t _ulticopy_65_start;
-extern uint32_t _ulticopy_65_size;
-
-#define CFG_IEC_ENABLE   0x51
-#define CFG_IEC_BUS_ID   0x52
-#define CFG_IEC_PATH     0x53
-
-static struct t_cfg_definition iec_config[] = {
-    { CFG_IEC_ENABLE,    CFG_TYPE_ENUM,   "IEC Drive",         "%s", en_dis, 0,  1, 0 },
-    { CFG_IEC_BUS_ID,    CFG_TYPE_VALUE,  "Soft Drive Bus ID", "%d", NULL,   8, 30, 11 },
-    { CFG_IEC_PATH,      CFG_TYPE_STRING, "Default Path",      "%s", NULL,   0, 30, (int) FS_ROOT },
-    { 0xFF, CFG_TYPE_END, "", "", NULL, 0, 0, 0 }
-};
-
-// this global will cause us to run!
-InitFunction iec_init(init_software_iec, NULL, NULL);
-void init_software_iec(void *_a, void *_b)
-{
-    iec_if = new IecInterface();
-    if (iec_if->get_current_iec_address() == 0) {
-        delete iec_if;
-        iec_if = NULL;
-    }
-}
-
-// Errors 
-
-// Error,ErrorText(,Track,Sector\015) 
-// The track and sector will be send seperately
-
-const char msg00[] = " OK";						//00
-const char msg01[] = " FILES SCRATCHED";		//01	Track number shows how many files were removed
-const char msg02[] = "PARTITION SELECTED";      //02
-const char msg20[] = "READ ERROR"; 				//20 (Block Header Not Found)
-//const char msg21[] = "READ ERROR"; 				//21 (No Sync Character)
-//const char msg22[] = "READ ERROR"; 				//22 (Data Block not Present)
-//const char msg23[] = "READ ERROR"; 				//23 (Checksum Error in Data Block)
-//const char msg24[] = "READ ERROR"; 				//24 (Byte Decoding Error)
-const char msg25[] = "WRITE ERROR";				//25 (Write/Verify Error)
-const char msg26[] = "WRITE PROTECT ON";		//26
-//const char msg27[] = "READ ERROR"; 				//27 (Checksum Error in Header)
-//const char msg28[] = "WRITE ERROR"; 			//28 (Long Data Block)
-const char msg29[] = "DISK ID MISMATCH";		//29
-const char msg30[] = "SYNTAX ERROR";			//30 (General)
-const char msg31[] = "SYNTAX ERROR";			//31 (Invalid Command)
-const char msg32[] = "SYNTAX ERROR";			//32 (Command Line > 58 Characters)
-const char msg33[] = "SYNTAX ERROR";			//33 (Invalid Filename)
-const char msg34[] = "SYNTAX ERROR";			//34 (No File Given)
-const char msg39[] = "SYNTAX ERROR";			//39 (Invalid Command)
-const char msg50[] = "RECORD NOT PRESENT";		//50
-const char msg51[] = "OVERFLOW IN RECORD";		//51
-//const char msg52[] = "FILE TOO LARGE";			//52
-const char msg60[] = "WRITE FILE OPEN";			//60
-const char msg61[] = "FILE NOT OPEN";			//61
-const char msg62[] = "FILE NOT FOUND";			//62
-const char msg63[] = "FILE EXISTS";				//63
-const char msg64[] = "FILE TYPE MISMATCH";		//64
-//const char msg65[] = "NO BLOCK";				//65
-//const char msg66[] = "ILLEGAL TRACK AND SECTOR";//66
-//const char msg67[] = "ILLEGAL SYSTEM T OR S";	//67
-const char msg69[] = "FILESYSTEM ERROR";        //69
-const char msg70[] = "NO CHANNEL";	            //70
-const char msg71[] = "DIRECTORY ERROR";			//71
-const char msg72[] = "DISK FULL";				//72
-const char msg73[] = "U64IEC ULTIMATE DOS V1.1";//73 DOS MISMATCH(Returns DOS Version)
-const char msg74[] = "DRIVE NOT READY";			//74
-const char msg77[] = "SELECTED PARTITION ILLEGAL"; //77
-const char msg_c1[] = "BAD COMMAND";			//custom
-const char msg_c2[] = "UNIMPLEMENTED";			//custom
-
-
-const IEC_ERROR_MSG last_error_msgs[] = {
-		{ 00,(char*)msg00,NR_OF_EL(msg00) - 1 },
-		{ 01,(char*)msg01,NR_OF_EL(msg01) - 1 },
-        { 02,(char*)msg02,NR_OF_EL(msg02) - 1 },
-		{ 20,(char*)msg20,NR_OF_EL(msg20) - 1 },
-//		{ 21,(char*)msg21,NR_OF_EL(msg21) - 1 },
-//		{ 22,(char*)msg22,NR_OF_EL(msg22) - 1 },
-//		{ 23,(char*)msg23,NR_OF_EL(msg23) - 1 },
-//		{ 24,(char*)msg24,NR_OF_EL(msg24) - 1 },
-		{ 25,(char*)msg25,NR_OF_EL(msg25) - 1 },
-		{ 26,(char*)msg26,NR_OF_EL(msg26) - 1 },
-//		{ 27,(char*)msg27,NR_OF_EL(msg27) - 1 },
-//		{ 28,(char*)msg28,NR_OF_EL(msg28) - 1 },
-		{ 29,(char*)msg29,NR_OF_EL(msg29) - 1 },
-		{ 30,(char*)msg30,NR_OF_EL(msg30) - 1 },
-		{ 31,(char*)msg31,NR_OF_EL(msg31) - 1 },
-		{ 32,(char*)msg32,NR_OF_EL(msg32) - 1 },
-		{ 33,(char*)msg33,NR_OF_EL(msg33) - 1 },
-		{ 34,(char*)msg34,NR_OF_EL(msg34) - 1 },
-		{ 39,(char*)msg39,NR_OF_EL(msg39) - 1 },
-		{ 50,(char*)msg50,NR_OF_EL(msg50) - 1 },
-		{ 51,(char*)msg51,NR_OF_EL(msg51) - 1 },
-//		{ 52,(char*)msg52,NR_OF_EL(msg52) - 1 },
-		{ 60,(char*)msg60,NR_OF_EL(msg60) - 1 },
-		{ 61,(char*)msg61,NR_OF_EL(msg61) - 1 },
-		{ 62,(char*)msg62,NR_OF_EL(msg62) - 1 },
-		{ 63,(char*)msg63,NR_OF_EL(msg63) - 1 },
-		{ 64,(char*)msg64,NR_OF_EL(msg64) - 1 },
-//		{ 65,(char*)msg65,NR_OF_EL(msg65) - 1 },
-//		{ 66,(char*)msg66,NR_OF_EL(msg66) - 1 },
-//		{ 67,(char*)msg67,NR_OF_EL(msg67) - 1 },
-        { 69,(char*)msg69,NR_OF_EL(msg69) - 1 },
-		{ 70,(char*)msg70,NR_OF_EL(msg70) - 1 },
-		{ 71,(char*)msg71,NR_OF_EL(msg71) - 1 },
-		{ 72,(char*)msg72,NR_OF_EL(msg72) - 1 },
-		{ 73,(char*)msg73,NR_OF_EL(msg73) - 1 },
-		{ 74,(char*)msg74,NR_OF_EL(msg74) - 1 },
-        { 77,(char*)msg77,NR_OF_EL(msg77) - 1 },
-
-		{ 75,(char*)msg_c1,NR_OF_EL(msg_c1) - 1 },
-		{ 76,(char*)msg_c2,NR_OF_EL(msg_c2) - 1 }
-};
-
-
-
-void IecInterface :: iec_task(void *a)
-{
-	IecInterface *iec = (IecInterface *)a;
-	iec->poll();
-}
-
-void IecInterface :: program_processor(void)
-{
-    printf("IEC Processor found: Version = %b. Loading code...", HW_IEC_VERSION);
-    HW_IEC_RESET_ENABLE = 0; // disable while programming
-    int size = (int)&_iec_code_b_size;
-    uint8_t *src = &_iec_code_b_start;
-    uint8_t *dst = (uint8_t *)HW_IEC_CODE;
-    for(int i=0; i<size; i++) {
-        *(dst++) = *(src++);
-    }
-}
-
-void IecInterface :: get_patch_locations(void)
-{
-    int size = (int)&_iec_code_b_size;
-    uint8_t *src = &_iec_code_b_start;
-
-    int found = 0;
-    for(int i=0; i<MAX_SLOTS; i++) {
-        talker_loc[i] = -1;
-        listener_loc[i] = -1;
-    }
-
-    for(int i=0; i<size; i+=4) {
-        // Each instruction is a 32 bit word 
-        uint32_t word_read = cpu_to_32le(src[i] | src[i+1]<<8 | src[i+2]<<16 | src[i+3]<<24);
-
-        int slot_id = word_read & 7;
-        // Find replacer words
-        if ((word_read & 0x1F8000F8) == 0x188000F0) {
-            if (slot_id < MAX_SLOTS) {
-                talker_loc[slot_id] = i;
-                // printf("Talker %d @ %d\n", slot_id, i / 4);
-                found++;
-            }
-        }
-        if ((word_read & 0x1F8000F8) == 0x188000F8) {
-            if (slot_id < MAX_SLOTS) {
-                listener_loc[slot_id] = i;
-                // printf("Listener %d @ %d\n", slot_id, i / 4);
-                found++;
-            }
-        }
-        if (found >= 2*MAX_SLOTS) {
-            break;
-        }
-    }
-    available_slots = found / 2;
-}
-
-void IecInterface :: iec_processor_configure(void)
-{
-    uint8_t drive_id = iec_drive_enabled ? iec_drive_id : 0x1f;
-    uint8_t printer_id = iec_printer_enabled ? iec_printer_id : 0x1f;
-
-    HW_IEC_RESET_ENABLE = 0; // disable
-    if (!iec_drive_enabled && !iec_printer_enabled) {
-        printf("Disable IEC processor\n");
-        return;
-    }
-
-    printf("Setting IEC bus ID to %d for drive and %d for printer.\n", iec_drive_id, iec_printer_id);
-
-    set_slot_devnum(0, drive_id);
-    set_slot_devnum(1, printer_id);
-    set_slot_devnum(2, 0x1f); // disable for now
-
-    HW_IEC_RESET_ENABLE = 1; // enable
-}
-
-void IecInterface :: set_slot_devnum(int slot, uint8_t dev)
-{
-    uint8_t *dst = (uint8_t *)HW_IEC_CODE;
-    dst[listener_loc[slot]] = dev | 0x20;
-    dst[talker_loc[slot]] = dev | 0x40;
-}
-
-void IecInterface :: iec_drive_enable(bool b, uint8_t id)
-{
-    iec_drive_enabled = b;
-    if (id) iec_drive_id = id;
-
-    iec_processor_configure();
-}
-
-void IecInterface :: iec_printer_enable(bool b, uint8_t id)
-{
-    iec_printer_enabled = b;
-    if (id) iec_printer_id = id;
-
-    iec_processor_configure();
-}
-
-IecInterface :: IecInterface() : SubSystem(SUBSYSID_IEC)
-{
-	fm = FileManager :: getFileManager();
-	ui_window = NULL;
-    last_addr = 0;
-    //return; // !!
-
-    if(!(getFpgaCapabilities() & CAPAB_HARDWARE_IEC))
-        return;
-
-    register_store(0x49454300, "Software IEC Settings", iec_config);
-
-    get_patch_locations();
-    program_processor();
-
-    iec_drive_enabled = 0;
-    iec_printer_enabled = 0;
-    iec_drive_id = 10;
-    iec_printer_id = 4;
-
-    atn = false;
-    cmd_path = fm->get_new_path("IEC Gui Path");
-    cmd_ui = 0;
-
-    last_error_code = ERR_DOS;
-    last_error_track = 0;
-    last_error_sector = 0;
-
-    current_channel = 0;
-    talking = false;
-    last_addr = 10;
-    wait_irq = false;
-    printer = false;
-
-    ulticopy_bin_image = new BinImage("UltiCopy", 35);
-
-    effectuate_settings();
-
-    vfs = new IecFileSystem(this);
-
-    for(int i=0;i<15;i++) {
-        channels[i] = new IecChannel(this, i);
-    }
-    channels[15] = new IecCommandChannel(this, 15);
- 
-    ulticopyBusy = xSemaphoreCreateBinary();
-    ulticopyMutex = xSemaphoreCreateMutex();
-    queueGuiToIec = xQueueCreate(2, sizeof(char *));
-
-    xTaskCreate( IecInterface :: iec_task, "IEC Server", configMINIMAL_STACK_SIZE, this, tskIDLE_PRIORITY + 2, &taskHandle );
-}
-
-IecInterface :: ~IecInterface()
-{
-    if(!(getFpgaCapabilities() & CAPAB_HARDWARE_IEC))
-        return;
-
-    if (taskHandle) {
-    	vTaskDelete(taskHandle);
-    }
-    for(int i=0;i<16;i++)
-        delete channels[i];
-
-    fm->release_path(cmd_path);
-}
-
-IecCommandChannel *IecInterface :: get_command_channel(void)
-{
-    return (IecCommandChannel *)channels[15];
-}
-
-IecCommandChannel *IecInterface :: get_data_channel(int chan)
-{
-    return (IecCommandChannel *)channels[chan & 15];
-}
-
-void IecInterface :: effectuate_settings(void)
-{
-    
-    int bus_id = cfg->get_value(CFG_IEC_BUS_ID);
-    cmd_if.set_kernal_device_id(bus_id);
-    
-    rootPath = cfg->get_string(CFG_IEC_PATH);
-    iec_enable = uint8_t(cfg->get_value(CFG_IEC_ENABLE));
-
-    iec_drive_enable(iec_enable, bus_id);
-}
-
-const char *IecInterface :: get_root_path(void)
-{
-    return rootPath;
-}
-
-void IecInterface :: create_task_items(void)
-{
-    TaskCategory *iec = TasksCollection :: getCategory("Software IEC", SORT_ORDER_SOFTIEC);
-    myActions.turn_on	 = new Action("Turn On",        SUBSYSID_IEC, MENU_IEC_ON);
-    myActions.reset      = new Action("Reset",          SUBSYSID_IEC, MENU_IEC_RESET);
-    myActions.set_dir    = new Action("Set dir. here",  SUBSYSID_IEC, MENU_IEC_SET_DIR);
-    myActions.turn_off	 = new Action("Turn Off",       SUBSYSID_IEC, MENU_IEC_OFF);
-    myActions.ulticopy8  = new Action("UltiCopy 8",     SUBSYSID_IEC, MENU_IEC_WARP_8);
-    myActions.ulticopy9  = new Action("UltiCopy 9",     SUBSYSID_IEC, MENU_IEC_WARP_9);
-    myActions.ulticopy10 = new Action("UltiCopy 10",    SUBSYSID_IEC, MENU_IEC_WARP_10);
-    myActions.ulticopy11 = new Action("UltiCopy 11",    SUBSYSID_IEC, MENU_IEC_WARP_11);
-
-    iec->append(myActions.turn_on);
-    iec->append(myActions.turn_off);
-    iec->append(myActions.reset);
-    iec->append(myActions.set_dir);
-    iec->append(myActions.ulticopy8);
-    iec->append(myActions.ulticopy9);
-    iec->append(myActions.ulticopy10);
-    iec->append(myActions.ulticopy11);
-}
-
-// called from GUI task
-void IecInterface :: update_task_items(bool writablePath, Path *path)
-{
-	if (iec_enable) {
-		myActions.turn_off->show();
-		myActions.turn_on->hide();
-	} else {
-		myActions.turn_on->show();
-		myActions.turn_off->hide();
-	}
-	if (writablePath) {
-		myActions.ulticopy8->enable();
-		myActions.ulticopy9->enable();
-		myActions.ulticopy10->enable();
-		myActions.ulticopy11->enable();
-	} else {
-		myActions.ulticopy8->disable();
-		myActions.ulticopy9->disable();
-		myActions.ulticopy10->disable();
-		myActions.ulticopy11->disable();
-	}
-}
-
-// this is actually the task
-void IecInterface :: poll()
-{
-    uint8_t data;
-    char *pathstring;
-    BaseType_t gotSomething;
-
-    while(1) {
-        if(wait_irq) {
-            if (HW_IEC_IRQ & 0x01) {
-                get_warp_data();
-            }
-            continue;
-        }
-
-        gotSomething = xQueueReceive(queueGuiToIec, &pathstring, 2); // here is the vTaskDelay(2) that used to be here
-
-        if (gotSomething == pdTRUE) {
-            if (!pathstring) {
-                start_warp_iec();
-            } else {
-                IecPartition *p = vfs->GetPartition(0);
-                p->cd(pathstring);
-                delete[] pathstring;
-            }
-        }
-
-        uint8_t a;
-        for (int loopcnt = 0; loopcnt < 500; loopcnt++) {
-            a = HW_IEC_RX_FIFO_STATUS;
-            if (a & IEC_FIFO_EMPTY) {
-                break;
-            }
-
-            data = HW_IEC_RX_DATA;
-
-            if(a & IEC_FIFO_CTRL) {
-                switch(data) {
-                    case WARP_START_RX:
-                        DBGIEC("[WARP:START.");
-                        HW_IEC_TX_DATA = 0x00; // handshake and wait for IRQ
-                        wait_irq = true;
-                        break;
-
-                    case WARP_BLOCK_END:
-                        DBGIEC(".WARP:OK]");
-                        ioWrite8(UART_DATA, 0x23); // printf("{warp_end}");
-                        break;
-
-                    case WARP_RX_ERROR:
-                        DBGIEC(".WARP:ERR]");
-                        get_warp_error();
-                        break;
-
-                    case WARP_ACK:
-                        DBGIEC(".WARP:ACK.");
-                        break;
-
-                    case CTRL_READY_FOR_TX:
-                        DBGIEC(".RTS.");
-                        channels[current_channel]->talk();
-                        HW_IEC_TX_FIFO_RELEASE = 1;
-                        talking = true;
-                        break;
-
-                    case CTRL_EOI:
-                        DBGIEC(".EOI]\n");
-                        if (printer) {
-                            iec_printer->push_command(0xFF);
-                            printer = false;
-                        } else {
-                            channels[current_channel]->push_command(0);
-                        }
-                        break;
-
-                    case CTRL_ATN_BEGIN:
-                        DBGIEC("\n[ATN.");
-                        atn = true;
-                        talking = false;
-                        printer = false;
-                        break;
-
-                    case CTRL_ATN_END:
-                        DBGIEC(".ATN>");
-                        atn = false;
-                        break;
-
-                    case CTRL_BYTE_TXED:
-                        DBGIEC("o");
-                        if (!printer) {
-                            channels[current_channel]->pop_data();
-                        }
-                        break;
-
-                    case CTRL_DEV2:
-                        DBGIEC(".PRINTER.");
-                        printer = true;
-                        iec_printer->push_command(0xFE);
-                        break;
-
-                    default:
-                        DBGIECV(".<%b>.", data);
-                        break;
-                }
-            } else {
-                if(atn) {
-                    DBGIECV(".<%b>.", data);
-
-                    if(data >= 0x60) {  // workaround for passing of wrong atn codes talk/untalk
-                        if (printer) {
-                            DBGIECV(".PRTCMD(%b).", data);
-                            iec_printer->push_command(data & 0x7);
-                        } else {
-                            DBGIECV(".DRVCMD(%b).", data);
-                            current_channel = int(data & 0x0F);
-                            channels[current_channel]->push_command(data & 0xF0);
-                        }
-                    }
-                } else {
-                    DBGIECV(".%b.", data);
-
-                    if (printer) {
-                        iec_printer->push_data(data);
-                    } else {
-                        channels[current_channel]->push_data(data);
-                    }
-                }
-            }
-
-            if (wait_irq) {
-                break;
-            }
-        }
-
-        int st;
-
-        if(talking) {
-            while(!(HW_IEC_TX_FIFO_STATUS & IEC_FIFO_FULL)) {
-                st = channels[current_channel]->prefetch_data(data);
-
-                if(st == IEC_OK) {
-                    HW_IEC_TX_DATA = data;
-#if IECDEBUG > 2
-                    outbyte(data < 0x20?'.':data);
-#endif
-                } else if(st == IEC_LAST) {
-                                        HW_IEC_TX_LAST = data;
-#if IECDEBUG > 2
-                    outbyte(data < 0x20?'.':data);
-#endif
-                    talking = false;
-                    break;
-                } else if(st == IEC_BUFFER_END) {
-                    break;
-                } else {
-                    printf("Talk Error = %d\n", st);
-                    HW_IEC_TX_CTRL = IEC_CMD_TX_TERM;
-                    talking = false;
-                    break;
-                }
-            }
-#if IECDEBUG > 2
-            outbyte('\'');
-#endif
-        }
-    }
-}
-
-// called from GUI task
-SubsysResultCode_e IecInterface :: executeCommand(SubsysCommand *cmd)
-{
-	File *f = 0;
-	uint32_t transferred;
-    char buffer[24];
-    int res;
-    FRESULT fres;
-
-    cmd_path->cd(cmd->path.c_str());
-    cmd_ui = cmd->user_interface;
-
-	switch(cmd->functionID) {
-		case MENU_IEC_ON:
-			iec_enable = 1;
-			cfg->set_value(CFG_IEC_ENABLE, iec_enable);
-                        iec_drive_enable(iec_enable);
-			break;
-		case MENU_IEC_OFF:
-			iec_enable = 0;
-			cfg->set_value(CFG_IEC_ENABLE, iec_enable);
-			iec_drive_enable(iec_enable);
-			break;
-		case MENU_IEC_RESET:
-            reset();
-			break;
-		case MENU_IEC_WARP_8:
-			start_warp(8);
-			break;
-		case MENU_IEC_WARP_9:
-			start_warp(9);
-			break;
-        case MENU_IEC_WARP_10:
-            start_warp(10);
-            break;
-        case MENU_IEC_WARP_11:
-            start_warp(11);
-            break;
-		case MENU_IEC_SET_DIR:
-		    set_iec_dir(cmd->path.c_str());
-		    break;
-		case MENU_IEC_MASTER_1:
-			test_master(1);
-			break;
-		case MENU_IEC_MASTER_2:
-			test_master(2);
-			break;
-		case MENU_IEC_MASTER_3:
-			test_master(3);
-			break;
-		case MENU_IEC_MASTER_4:
-			test_master(4);
-			break;
-		case MENU_IEC_LOADDIR:
-			master_open_file(8, 0, "$", false);
-			break;
-		case MENU_IEC_LOADFIRST:
-			master_open_file(8, 0, "*", false);
-			break;
-		case MENU_READ_STATUS:
-			master_read_status(8);
-			break;
-		case MENU_SEND_COMMAND:
-			buffer[0] = 0;
-			res = cmd->user_interface->string_box("Command", buffer, 22);
-			if (res > 0) {
-				master_send_cmd(8, (uint8_t*)buffer, strlen(buffer));
-			}
-			break;
-		default:
-			break;
-    }
-    return SSRET_OK;
-}
-
-void IecInterface :: reset(void)
-{
-    HW_IEC_RESET_ENABLE = 0;
-    for(int i=0; i < 16; i++) {
-        channels[i]->reset();
-    }
-    iec_drive_enable(iec_enable);
-}
-
-#include "c1541.h"
-extern C1541 *c1541_A;
-extern C1541 *c1541_B;
-
-// called from GUI task
-void IecInterface :: set_iec_dir(const char *path)
-{
-    char *pathcopy = new char[strlen(path) + 1];
-    strcpy(pathcopy, path);
-    xQueueSend(queueGuiToIec, &pathcopy, 0); // write into command queue
-}
-
-// called from critical section
-const char *IecInterface :: get_partition_dir(int p)
-{
-    return vfs->GetPartitionPath(p);
-}
-
-// called from GUI task
-void IecInterface :: start_warp(int drive)
-{
-	// First try to obtain a lock on this function
-	if (xSemaphoreTake(ulticopyMutex, 0) == pdFALSE) {
-		cmd_ui->popup("Ulticopy is already in use", BUTTON_OK);
-		return;
-	}
-
-	// FIXME: Make sure that the C64 is actually frozen!
-    printf("Starting IEC Warp Mode.\n");
-    C64_POKE(0xDD00,0x07); // make sure the C64 that might be connected does not interfere
-    // make sure that our local drives are turned off as well
-    if (c1541_A) {
-        c1541_A->drive_power(false);
-    }
-    if (c1541_B) {
-        c1541_B->drive_power(false);
-    }
-
-    ui_window = new UltiCopy();
-    ui_window->init(cmd_ui->screen, cmd_ui->keyboard);
-    cmd_ui->activate_uiobject(ui_window); // now we have focus
-    ui_window->window->move_cursor(15,10);
-    ui_window->window->output("Loading...");
-
-    warp_drive = drive;
-    const char *dummy = NULL;
-    xQueueSend(queueGuiToIec, &dummy, 0); // ulticopy shall now take over
-
-    if (xSemaphoreTake(ulticopyBusy, 10) == pdFALSE) {
-    	printf("Synchronization error!  UltiCopy did not start.\n");
-    	while(1)
-    		;
-    }
-    while(1) {
-		printf("GUI Task was notified that UltiCopy is now busy");
-		xSemaphoreTake(ulticopyBusy, portMAX_DELAY);
-
-		// Ulticopy is done
-		printf("UltiCopy warp mode returned.\n");
-
-		if(warp_return_code & 0x80) {
-			last_track++;
-			printf("Error on track %d.\n", last_track);
-		} else if(warp_return_code == 0) {
-			save_copied_disk();
-			if(cmd_ui->popup("Another disk?", BUTTON_YES|BUTTON_NO) == BUTTON_YES) {
-				ui_window->window->clear();
-				// if(!run_drive_code(warp_drive, 0x403, NULL, 0)) { // restart
-			    if(!run_drive_code(warp_drive, 0x400, &_ulticopy_65_start, (int)&_ulticopy_65_size)) {
-					cmd_ui->popup("Restart error..", BUTTON_OK);
-					break;
-			    }
-				// clear pending interrupt if any
-				HW_IEC_IRQ = 0;
-				// push warp command into down-fifo
-				HW_IEC_TX_FIFO_RELEASE = 1;
-				HW_IEC_TX_CTRL = IEC_CMD_GO_WARP;
-				last_track = 0;
-			} else {
-				ui_window->close();
-				break;
-			}
-		} else if(warp_return_code < 0x40) {
-			cmd_ui->popup("Error reading disk..", BUTTON_OK);
-			ui_window->close();
-			break;
-		}
-	}
-	if (c1541_A) {
-		c1541_A->effectuate_settings();
-	}
-	if (c1541_B) {
-		c1541_B->effectuate_settings();
-	}
-	HW_IEC_RESET_ENABLE = iec_enable;
-
-    xSemaphoreGive(ulticopyMutex);
-}
-
-void IecInterface :: start_warp_iec(void)
-{
-	// notify that we are busy now
-	xSemaphoreGive(ulticopyBusy);
-
-    HW_IEC_RESET_ENABLE = 1;
-    
-    if(!run_drive_code(warp_drive, 0x400, &_ulticopy_65_start, (int)&_ulticopy_65_size)) {
-        cmd_ui->popup("Error accessing drive..", BUTTON_OK);
-        ui_window->close();
-        warp_return_code = 0x1F;
-        xSemaphoreGive(ulticopyBusy);
-        return;
-    }
-
-    ui_window->window->clear();
-    // clear pending interrupt if any
-    HW_IEC_IRQ = 0;
-    // push warp command into down-fifo; Go!
-	HW_IEC_TX_FIFO_RELEASE = 1;
-    HW_IEC_TX_CTRL = IEC_CMD_GO_WARP;
-    last_track = 0;
-}
-
-void IecInterface :: get_warp_data(void)
-{
-    uint32_t read;
-    uint32_t temp[65];
-    uint32_t *dw = &temp[0];
-
-    int err = 0;
-    uint16_t fifo_count = uint16_t(HW_IEC_UP_FIFO_COUNT_LO) | (uint16_t(HW_IEC_UP_FIFO_COUNT_HI) << 8);
-
-    for(int i=0;i<64;i++) {
-    	GCR_DECODER_GCR_IN_32 = HW_IEC_RX_DATA_32; // first in first out, endianness OK
-        GCR_DECODER_GCR_IN = HW_IEC_RX_DATA;
-        *(dw++) = GCR_DECODER_BIN_OUT_32;
-        if(GCR_DECODER_ERRORS)
-            err++;
-    }
-    read = HW_IEC_RX_DATA_32;
-    GCR_DECODER_GCR_IN_32 = read;
-    GCR_DECODER_GCR_IN = 0x55;
-    *(dw++) = GCR_DECODER_BIN_OUT_32;
-
-    read = cpu_to_32le(read);
-    uint8_t sector = (uint8_t)(read >> 24);
-
-    uint8_t track = HW_IEC_RX_DATA;
-    uint8_t *dest = ulticopy_bin_image->get_sector_pointer(track, sector);
-    uint8_t *src = (uint8_t *)temp;
-    // printf("Sector {%b %b (%p -> %p}\n", track, sector, src, dest);
-    if (dest) {
-		ui_window->window->set_char(track-1,sector+1,err?'-':'*');
-		ui_window->parent_win->sync();
-		last_track = track;
-		if(dest) {
-			for(int i=0;i<256;i++) {
-				*(dest++) = *(++src); // asymmetric: We copy from 1.. to 0..., so we increment src first
-			}
-		}
-    }
-    // clear pending interrupt
-    wait_irq = false;
-    HW_IEC_IRQ = 0;
-}
-
-// called from IEC context
-void IecInterface :: get_warp_error(void)
-{
-    while (HW_IEC_RX_FIFO_STATUS & IEC_FIFO_EMPTY)
-        ;
-        
-    warp_return_code = HW_IEC_RX_DATA;
-    printf("{Warp Error: %b}", warp_return_code);
-    // clear pending interrupt
-    HW_IEC_IRQ = 0;
-    // notify the gui that we are done
-    xSemaphoreGive(ulticopyBusy);
-}
-
-void IecInterface :: save_copied_disk()
-{
-    static char buffer[40];
-    int save_result;
-    File *f = 0;
-    int res;
-    BinImage *bin;
-    
-    ulticopy_bin_image->num_tracks = 35; // standard!
-
-	// buffer[0] = 0;
-    if (cmd_ui->cfg->get_value(CFG_USERIF_ULTICOPY_NAME)) {
-        ulticopy_bin_image->get_sensible_name(buffer);
-    }
-	res = cmd_ui->string_box("Give name for copied disk..", buffer, 22);
-	if(res > 0) {
-		fix_filename(buffer);
-		set_extension(buffer, ".d64", 32);
-        FRESULT fres = fm->fopen(cmd_path, buffer, FA_WRITE | FA_CREATE_NEW | FA_CREATE_ALWAYS, &f);
-		if(f) {
-            cmd_ui->show_progress("Saving D64..", 100);
-            save_result = ulticopy_bin_image->save(f, cmd_ui);
-            cmd_ui->hide_progress();
-    		printf("Result of save: %d.\n", save_result);
-            fm->fclose(f);
-		} else {
-			printf("Can't create file '%s': %s\n", buffer, FileSystem::get_error_string(fres));
-			cmd_ui->popup("Can't create file.", BUTTON_OK);
-		}
-	}
-}
-                
-void IecInterface :: set_error(int code, int track, int sector)
-{
-    last_error_code = code;
-    last_error_track = track;
-    last_error_sector = sector;
-}
-
-void IecInterface :: set_error_fres(FRESULT fres)
-{
-    int tr=0, sec=0, err = 0;
-
-    switch (fres) {
-    case FR_OK:                  err = ERR_ALL_OK;            /* (0) Succeeded */
-        break;
-    case FR_DISK_ERR:            err = ERR_DRIVE_NOT_READY;   /* (1) A hard error occurred in the low level disk I/O layer */
-        break;
-    case FR_INT_ERR:             err = ERR_FRESULT_CODE;      /* (2) Assertion failed */
-        break;
-    case FR_NOT_READY:           err = ERR_DRIVE_NOT_READY;   /* (3) The physical drive cannot work */
-        break;
-    case FR_NO_FILE:             err = ERR_FILE_NOT_FOUND;    /* (4) Could not find the file */
-        break;
-    case FR_NO_PATH:             err = ERR_FILE_NOT_FOUND;    /* (5) Could not find the path */
-        break;
-    case FR_INVALID_NAME:        err = ERR_SYNTAX_ERROR_NAME; /* (6) The path name format is invalid */
-        break;
-    case FR_DENIED:              err = ERR_DIRECTORY_ERROR;   /* (7) Access denied due to prohibited access or directory full */
-        break;
-    case FR_EXIST:               err = ERR_FILE_EXISTS;       /* (8) Access denied due to prohibited access */
-        break;
-    case FR_INVALID_OBJECT:      err = ERR_FRESULT_CODE;      /* (9) The file/directory object is invalid */
-        break;
-    case FR_WRITE_PROTECTED:     err = ERR_WRITE_PROTECT_ON;  /* (10) The physical drive is write protected */
-        break;
-    case FR_INVALID_DRIVE:       err = ERR_DRIVE_NOT_READY;   /* (11) The logical drive number is invalid */
-        break;
-    case FR_NOT_ENABLED:         err = ERR_DRIVE_NOT_READY;   /* (12) The volume has no work area */
-        break;
-    case FR_NO_FILESYSTEM:       err = ERR_DRIVE_NOT_READY;   /* (13) There is no valid FAT volume */
-        break;
-    case FR_MKFS_ABORTED:        err = ERR_FRESULT_CODE;      /* (14) The f_mkfs() aborted due to any parameter error */
-        break;
-    case FR_TIMEOUT:             err = ERR_DRIVE_NOT_READY;   /* (15) Could not get a grant to access the volume within defined period */
-        break;
-    case FR_LOCKED:              err = ERR_WRITE_FILE_OPEN;   /* (16) The operation is rejected according to the file sharing policy */
-        break;
-    case FR_NO_MEMORY:           err = ERR_FRESULT_CODE;      /* (17) LFN working buffer could not be allocated */
-        break;
-    case FR_TOO_MANY_OPEN_FILES: err = ERR_FRESULT_CODE;      /* (18) Number of open files > _FS_SHARE */
-        break;
-    case FR_INVALID_PARAMETER:   err = ERR_FRESULT_CODE;      /* (19) Given parameter is invalid */
-        break;
-    case FR_DISK_FULL:           err = ERR_DISK_FULL;         /* (20) OLD FATFS: no more free clusters */
-        break;
-    case FR_DIR_NOT_EMPTY:       err = ERR_FILE_EXISTS;       /* (21) Directory not empty */
-        break;
-    default:                     err = ERR_FRESULT_CODE;
-    }
-    if (err == ERR_FRESULT_CODE) {
-        tr = fres;
-    }
-    set_error(err, tr, sec);
-}
-
-int IecInterface ::get_error_string(char *buffer)
-{
-    int len;
-    for (int i = 0; i < NR_OF_EL(last_error_msgs); i++) {
-        if (last_error_code == last_error_msgs[i].nr) {
-            return sprintf(buffer, "%02d,%s,%02d,%02d\015", last_error_code, last_error_msgs[i].msg,
-                            last_error_track, last_error_sector);
-        }
-    }
-    return sprintf(buffer, "99,UNKNOWN,00,00\015");
-}
-
-void IecInterface :: master_open_file(int device, int channel, const char *filename, bool write)
-{
-    printf("Open '%s' on device '%d', channel %d\n", filename, device, channel);
-	HW_IEC_TX_FIFO_RELEASE = 1;
-    HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-    HW_IEC_TX_DATA = 0x20 | (((uint8_t)device) & 0x1F); // listen!
-    HW_IEC_TX_DATA = 0xF0 | (((uint8_t)channel) & 0x0F); // open on channel x
-    HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_TX;
-
-    int len = strlen(filename);
-    for (int i=0;i<len;i++) {
-        while(HW_IEC_TX_FIFO_STATUS & IEC_FIFO_FULL)
-            ;
-        if (i == (len-1)) {
-            HW_IEC_TX_LAST = (uint8_t)filename[i]; // EOI
-        } else {
-        	HW_IEC_TX_DATA = (uint8_t)filename[i];
-        }
-    }
-    if(!write) {
-        while(!(HW_IEC_TX_FIFO_STATUS & IEC_FIFO_EMPTY))
-            ;
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x3F; // unlisten
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_RELEASE;
-        // and talk!
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x40 | (((uint8_t)device) & 0x1F); // talk
-        HW_IEC_TX_DATA = 0x60 | (((uint8_t)channel) & 0x0F); // open channel
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_RX;
-    }
-}        
-
-bool IecInterface :: master_send_cmd(int device, uint8_t *cmd, int length)
-{
-    //printf("Send command on device '%d' [%s]\n", device, cmd);
-    // dump_hex(cmd, length);
-    
-	HW_IEC_TX_FIFO_RELEASE = 1;
-    HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-    HW_IEC_TX_DATA = 0x20 | (((uint8_t)device) & 0x1F); // listen!
-    HW_IEC_TX_DATA = 0x6F;
-    HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_TX;
-
-    for (int i=0;i<length;i++) {
-        while(HW_IEC_TX_FIFO_STATUS & IEC_FIFO_FULL)
-            ;
-        if (i == (length-1)) {
-            HW_IEC_TX_LAST = cmd[i];
-        } else {
-            HW_IEC_TX_DATA = cmd[i];
-        }
-    }
-    while(!(HW_IEC_TX_FIFO_STATUS & IEC_FIFO_EMPTY))
-        ;
-
-    HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-    HW_IEC_TX_DATA = 0x3F; // unlisten
-    HW_IEC_TX_CTRL = IEC_CMD_ATN_RELEASE;
-
-    while(!(HW_IEC_TX_FIFO_STATUS & IEC_FIFO_EMPTY))
-        ;
-
-    uint8_t st,code;
-    st = HW_IEC_RX_FIFO_STATUS;
-    if(!(st & IEC_FIFO_EMPTY)) {
-        code = HW_IEC_RX_DATA;
-        if(st & IEC_FIFO_CTRL) {
-            printf("Return code: %b\n", code);
-            if((code & 0xF0) == 0xE0)
-                return false;
-        } else {
-            printf("Huh? Didn't expect data: %b\n", code);
-        }        
-    }
-    return true;
-}
-
-void IecInterface :: master_read_status(int device)
-{
-    printf("Reading status channel from device %d\n", device);
-	HW_IEC_TX_FIFO_RELEASE = 1;
-    HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-    HW_IEC_TX_DATA = 0x40 | (((uint8_t)device) & 0x1F); // listen!
-    HW_IEC_TX_DATA = 0x6F; // channel 15
-    HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_RX;
-}
-
-bool IecInterface :: run_drive_code(int device, uint16_t addr, uint8_t *code, int length)
-{
-    printf("Load drive code. Length = %d\n", length);
-    uint8_t buffer[40];
-    uint16_t address = addr;
-    int size;
-    strcpy((char*)buffer, "M-W");
-    while(length > 0) {
-        size = (length > 32)?32:length;
-        buffer[3] = (uint8_t)(address & 0xFF);
-        buffer[4] = (uint8_t)(address >> 8);
-        buffer[5] = (uint8_t)(size);
-        for(int i=0;i<size;i++) {
-            buffer[6+i] = *(code++);
-        }
-        printf(".");
-        if(!master_send_cmd(device, buffer, size+6))
-            return false;
-        length -= size;
-        address += size;
-    }
-    strcpy((char*)buffer, "M-E");
-    buffer[3] = (uint8_t)(addr & 0xFF);
-    buffer[4] = (uint8_t)(addr >> 8);
-    return master_send_cmd(device, buffer, 5);
-}
-
-void IecInterface :: test_master(int test)
-{
-    switch(test) {
-    case 1:
-        printf("Open $ on device 8\n");
-		HW_IEC_TX_FIFO_RELEASE = 1;
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x28; // listen #8!
-        HW_IEC_TX_DATA = 0xF0; // open
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_TX;
-        HW_IEC_TX_LAST = '$';
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x3F; // unlisten
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_RELEASE;
-        break;
-    case 2:
-        printf("Talk channel 0\n");
-		HW_IEC_TX_FIFO_RELEASE = 1;
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x48; // talk #8!!
-        HW_IEC_TX_DATA = 0x60; // channel 0
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_RX;
-        break;
-    case 3:
-        printf("Close file on 8\n");
-		HW_IEC_TX_FIFO_RELEASE = 1;
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x5F; // untalk
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_RELEASE;
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x28; // listen #8!
-        HW_IEC_TX_DATA = 0xE0; // close
-        HW_IEC_TX_DATA = 0x3F; // unlisten
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_RELEASE;
-        break;
-    case 4:
-        printf("Reading status channel\n");
-		HW_IEC_TX_FIFO_RELEASE = 1;
-        HW_IEC_TX_CTRL = IEC_CMD_GO_MASTER;
-        HW_IEC_TX_DATA = 0x48; // talk #8!!
-        HW_IEC_TX_DATA = 0x6F; // channel 15
-        HW_IEC_TX_CTRL = IEC_CMD_ATN_TO_RX;
-        break;        
-    default:
-        printf("To be defined.\n");
-    }
-}
-
-/*********************************************************************/
-/* Copier user interface screen
-/*********************************************************************/
-UltiCopy :: UltiCopy()
-{
-    parent_win = NULL;
-    keyb = NULL;
-    window = NULL;
-    return_code = 0;
-}
-    
-UltiCopy :: ~UltiCopy()
-{
-}    
-
-void UltiCopy :: init(Screen *scr, Keyboard *key)
-{
-    parent_win = scr;
-    keyb = key;
-    window = new Window(parent_win, 0, 1, 40, 23);
-    window->draw_border_horiz();
-    window->clear();
-}
-    
-void UltiCopy :: deinit(void)
-{
-    if (window)
-        delete window;
-}
-
-int UltiCopy :: handle_key(uint8_t c)
-{
-	return 0;
-}
-
-int UltiCopy :: poll(int a)
-{
-    return return_code;
-}
-        
-void UltiCopy :: close(void)
-{
-    return_code = 1;
-}
-    
-void iec_info(StreamTextLog &b)
-{
-    char buffer[64];
-    if (!iec_if) {
-        return;
-    }
-    b.format("SoftwareIEC:  %s\n", iec_if->iec_enable ? "Enabled" : "Disabled");
-    if (iec_if->iec_enable) {
-        b.format("Drive Bus ID: %d\n", iec_if->get_current_iec_address());
-        iec_if->get_error_string(buffer);
-        b.format("Error string: %s\n", buffer);
-        for (int i=0; i < MAX_PARTITIONS; i++) {
-            const char *p = iec_if->get_partition_dir(i);
-            if (p) {
-                b.format("Partition%3d: %s\n", i, p);
-            }
-        }
-    }
-    b.format("\n");
-}
-
-void iec_info(JSON_List *obj)
-{
-    char buffer[64];
-    if (!iec_if) {
-        return;
-    }
-    iec_if->get_error_string(buffer);
-    
-    JSON_List *partitions = JSON::List();
-    for (int i=0; i < MAX_PARTITIONS; i++) {
-        const char *p = iec_if->get_partition_dir(i);
-        if (p) {
-            partitions->add(JSON::Obj()->add("id", i)->add("path", p));
-        }
-    }
-
-    obj->add(JSON::Obj()->add("softiec", JSON::Obj()
-        ->add("enabled", (bool)iec_if->iec_enable)
-        ->add("bus_id", iec_if->get_current_iec_address())
-        ->add("type", "DOS emulation")
-        ->add("last_error", buffer)
-        ->add("partitions", partitions)));
-}
-
-
-/*********************************************************************/
-/* IEC File Browser Handling                                         */
-/*********************************************************************/
-/*
-#define IECFILE_LOAD 0xCA01
-
-FileTypeIEC tester_iec(file_type_factory);
-
-FileTypeIEC :: FileTypeIEC(CachedTreeNode *par, FileInfo *fi) : FileDirEntry(par, fi)
-{
-}
-
-FileTypeIEC :: ~FileTypeIEC()
-{
-}
-
-int   FileTypeIEC :: fetch_context_items(IndexedList<Action *> &list)
-{
-	list.append(new MenuItem(this, "Load Code", IECFILE_LOAD));
-    return 1+ (FileDirEntry :: fetch_context_items_actual(list));
-}
-
-FileDirEntry *FileTypeIEC :: test_type(CachedTreeNode *obj)
-{
-	FileInfo *inf = obj->get_file_info();
-    if(strcmp(inf->extension, "IEC")==0)
-        return new FileTypeIEC(obj->parent, inf);
-    return NULL;
-}
-
-void FileTypeIEC :: execute(int selection)
-{
-    File *file;
-    uint32_t bytes_read;
-    
-	switch(selection) {
-        case IECFILE_LOAD:
-            printf("Loading IEC file.\n");
-    		file = root.fopen(this, FA_READ);
-    		if(file) {
-                HW_IEC_RESET_ENABLE = 0;
-                file->read((void *)HW_IEC_CODE, 2048, &bytes_read);
-                printf("Read %d code bytes.\n", bytes_read);
-                HW_IEC_RESET_ENABLE = iec_if.iec_enable;
-            }
-            break;
-        default:
-            break;
-    }
-}
-*/
-
